+++ server/apps/models/priv/repo/migrations/20170319055036_create_combat_lifetime_statistics.exs
@@ -22,14 +22,14 @@ defmodule Models.Repo.Migrations.CreateCombatLifetimeStatistics do
       add :shield_generators_destroyed, :integer, null: false, default: 0
       add :damage_blocked, :integer, null: false, default: 0
       add :melee_kills, :integer, null: false, default: 0
-      add :weapon_accuracy_percentage, :integer, null: false, default: 0
-      add :critical_hits_accuracy_percentage, :integer, null: false, default: 0
+      add :weapon_accuracy_percentage, :decimal, null: false, default: 0
+      add :critical_hits_accuracy_percentage, :decimal, null: false, default: 0
       add :turrets_destroyed, :integer, null: false, default: 0
       add :defensive_assists, :integer, null: false, default: 0
       add :offensive_assists, :integer, null: false, default: 0
       add :barrier_damage_done, :integer, null: false, default: 0
       add :hero_damage_done, :integer, null: false, default: 0
-      add :scoped_accuracy_percentage, :integer, null: false, default: 0
+      add :scoped_accuracy_percentage, :decimal, null: false, default: 0
       add :self_healing, :integer, null: false, default: 0
       add :damage_amplified, :integer, null: false, default: 0
       add :critical_hits_kills, :integer, null: false, default: 0
